Fix warnings in all renders

+++ src/Layers/xrRender/FSkinned.cpp
@@ -608,7 +608,7 @@ void CSkeletonX_ext::_Load_hw	(Fvisual& V, void *	_verts_)
 	case RM_SINGLE:
 	case RM_SKINNING_1B:
 		{
-			u32 vStride = ComputeVertexSize(dwDecl_01W, 0);
+			u32 vStride = (u32)ComputeVertexSize(dwDecl_01W, 0);
 			VERIFY	(vStride==sizeof(vertHW_1W));
 			BYTE*	bytes		= 0;
 			VERIFY				(NULL==V.p_rm_Vertices);
@@ -628,7 +628,7 @@ void CSkeletonX_ext::_Load_hw	(Fvisual& V, void *	_verts_)
 		break;
 	case RM_SKINNING_2B:
 		{
-			u32 vStride = ComputeVertexSize(dwDecl_2W, 0);
+			u32 vStride = (u32)ComputeVertexSize(dwDecl_2W, 0);
 			VERIFY				(vStride==sizeof(vertHW_2W));
 			BYTE* bytes			= 0;
 			VERIFY				(NULL==V.p_rm_Vertices);
@@ -649,7 +649,7 @@ void CSkeletonX_ext::_Load_hw	(Fvisual& V, void *	_verts_)
 		}break;
 	case RM_SKINNING_3B:
 		{
-			u32 vStride = ComputeVertexSize(dwDecl_3W, 0);
+			u32 vStride = (u32)ComputeVertexSize(dwDecl_3W, 0);
 			VERIFY					(vStride==sizeof(vertHW_3W));
 			BYTE*	bytes			= 0;
 			VERIFY					(NULL==V.p_rm_Vertices);
@@ -671,7 +671,7 @@ void CSkeletonX_ext::_Load_hw	(Fvisual& V, void *	_verts_)
 		}break;
 	case RM_SKINNING_4B:
 		{
-			u32 vStride = ComputeVertexSize(dwDecl_4W, 0);
+			u32 vStride = (u32)ComputeVertexSize(dwDecl_4W, 0);
 			VERIFY					(vStride==sizeof(vertHW_4W));
 			BYTE*	bytes			= 0;
 			VERIFY					(NULL==V.p_rm_Vertices);
