+++ src/Layers/xrRenderPC_R2/light_smapvis.cpp
@@ -107,7 +107,7 @@ void	smapvis::resetoccq	()
 
 void	smapvis::mark				()
 {
-	RImplementation.stats.ic_culled	+= invisible.size	();
+	RImplementation.stats.ic_culled	+= (u32)invisible.size	();
 	u32		marker			= RImplementation.marker + 1;	// we are called befor marker increment
 	for		(u32 it=0; it<invisible.size(); it++)
 		invisible[it]->vis.marker	= marker;				// this effectively disables processing
